mentor's fix (Invalid ID message)

+++ controllers/contactsControllers.js
@@ -1,3 +1,4 @@
+import mongoose from "mongoose";
 import HttpError from "../helpers/HttpError.js";
 import {
   createContactSchema,
@@ -5,6 +6,8 @@ import {
   updateStatusSchema,
 } from "../schemas/contactsSchemas.js";
 import * as contactsService from "../services/contactsServices.js";
+
+const { isValidObjectId } = mongoose;
 
 export const getAllContacts = async (req, res, next) => {
   try {
@@ -18,6 +21,9 @@ export const getAllContacts = async (req, res, next) => {
 export const getOneContact = async (req, res, next) => {
   try {
     const { id } = req.params;
+    if (!isValidObjectId(id)) {
+      throw HttpError(400, "Invalid ID format");
+    }
     const result = await contactsService.getContactById(id);
     if (!result) {
       throw HttpError(404);
@@ -31,6 +37,10 @@ export const getOneContact = async (req, res, next) => {
 export const deleteContact = async (req, res, next) => {
   try {
     const { id } = req.params;
+
+    if (!isValidObjectId(id)) {
+      throw HttpError(400, "Invalid ID format");
+    }
     const result = await contactsService.removeContact(id);
     if (!result) {
       throw HttpError(404);
@@ -57,13 +67,21 @@ export const createContact = async (req, res, next) => {
 export const updateContact = async (req, res, next) => {
   try {
     const { id } = req.params;
+
+    if (!isValidObjectId(id)) {
+      throw HttpError(400, "Invalid ID format");
+    }
+
     const { error } = updateContactSchema.validate(req.body);
     if (error) throw HttpError(400, error.message);
 
     if (Object.keys(req.body).length === 0) {
       throw HttpError(400, "Body must have at least one field");
     }
-    const result = await contactsService.updateContact(id, req.body);
+
+    const { favorite, ...updateBody } = req.body;
+
+    const result = await contactsService.updateContact(id, updateBody);
     if (!result) {
       throw HttpError(404);
     }
@@ -77,6 +95,9 @@ export const updateContact = async (req, res, next) => {
 export const updateStatus = async (req, res, next) => {
   try {
     const { id } = req.params;
+    if (!isValidObjectId(id)) {
+      throw HttpError(400, "Invalid ID format");
+    }
     const { error } = updateStatusSchema.validate(req.body);
     if (error) throw HttpError(400, error.message);
